Implement ViewSuggestionsCommand

+++ src/main/java/fitnus/command/GenerateCalorieGoalCommand.java
@@ -29,7 +29,7 @@ public class GenerateCalorieGoalCommand extends Command {
                     + "1 kg. Please try again with a lower weekly goal!";
         } else {
             int bmr; //basal metabolic rate
-            int calDiff = (int) Math.round(weeklyChange * calDeficitFor1KgWeekly);
+            int calDiff = Math.round(weeklyChange * calDeficitFor1KgWeekly);
             int newGoal;
             if (us.getGender() == MALE) {
                 bmr = (int) Math.round(((655.1 + (9.563 * us.getWeight())
